+++ vpc/nat_gateway/variables.tf
@@ -5,6 +5,15 @@ variable "nat_map" {
   }))
 }
 
+variable "std_map" {
+  type = object({
+    name_replace_regex   = string
+    resource_name_prefix = string
+    resource_name_suffix = string
+    tags                 = map(string)
+  })
+}
+
 variable "subnet_id_map" {
   type        = map(string)
   description = "Map of k_az_full to subnet ID"